backend/requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.16,<3.17
django-cors-headers>=4.4,<5.0
drf-spectacular>=0.28,<0.29
psycopg2-binary>=2.9,<3.0
celery>=5.4,<6.0
redis>=5.0,<6.0
